% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyvarianceFilter.R
\name{applyvarianceFilter}
\alias{applyvarianceFilter}
\title{Removes features that have near zero variance}
\usage{
applyvarianceFilter(data_object, threshold = 0.1)
}
\arguments{
\item{data_object}{argument is the output produced by as.ML function, which contains a single x data frame or a list of x data frames, a y data frames and attributes}

\item{threshold}{is the cutoff value for percentage of partitions containing features with near zero variance}
}
\description{
This function removes features that are above a percentage threshold of features with near zero variance
}
\details{

}
\examples{
dontrun{
}


}
