% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationFilter.R
\name{correlationFilter}
\alias{correlationFilter}
\title{Removes features that are clustered by correlation, replacing each cluster
with a single feature at a minimum within cluster correlation threshold}
\usage{
correlationFilter(data_object)
}
\arguments{
\item{data_object}{argument is the output produced by as.ML function, which
contains a single x data frame or a list of x data frames, a y data frames
and attributes}
}
\description{
This function removes features for a single data frame or a list of data
frames
}
\details{

}
