% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLSelection.R
\name{MLSelection}
\alias{MLSelection}
\title{Applies the MLWrapper to all data sources with all algorithms and returns
performance and timing metrics for each source/algorithm combination}
\usage{
MLSelection(data_object, partition_style = "random", folds = 3,
  repeats = 10, algorithms = "all", data_sources = "all",
  single_source = NULL)
}
\arguments{
\item{data_object}{argument is the output produced by as.MLinput, which
contains a single x data frame or a list of x data frames, a y data frames
and attributes}

\item{partition_style}{one of 'random' or 'paired' (character string)
indicating type of partition}

\item{folds}{integer of k for k-fold cross validation}

\item{repeats}{integer of number of iterations to repeat cross validation}

\item{algorithms}{either 'all', or one of 'rf', 'svm', 'nb', 'knn'. May also
include a vector combination of the latter (e.g. c('rf','svm'))}
}
\description{
importFrom magrittr "%>%"
}
