% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveIntegration.R
\name{naiveIntegration}
\alias{naiveIntegration}
\title{Combine probabilities from multiple data sources.}
\usage{
naiveIntegration(probabilities_by_source)
}
\arguments{
\item{probabilities_by_source}{a list of data.frames with one column for each
predicted class corresponding to the probability of that class label given
the data source. Each data frame corresponds to a data source. All
probabilities in each row must sum to 1. The rows are generally cross
validated test sets.}
}
\description{
Combine probabilities from multiple data sources.
}
\details{
This function combines class probabilities from learning algorithms
 on multiple or single data sources over cross validation partitions. The
 mechanism for combining the probabilities is a posterior probability where
 \deqn{P(c_{j} | x_{i}) = /prod{P(c_{j} | x_{i}^{s})}_{s=1}^{ns}} where
 \eqn{c} is one of \eqn{j} class labels, \eqn{x} is observed data for one of
 \eqn{ns} data sources.
}
