% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLmethods.R
\name{peppuR_knn}
\alias{peppuR_knn}
\title{KNN method}
\usage{
peppuR_knn(X_train, X_test, train_class, train_partition, test_partition,
  ...)
}
\arguments{
\item{...}{- additional arguements passed to BGLR(...)}

\item{X}{- n-by-p data.frame containing all of the 'covariates' that will be used to predict class of y}

\item{data}{- a data.frame with columns:
   y - n vector of class identifiers for each for of X
  ID - n vector of patient IDs
pair - vector of pair identifiers
time - n vector of time point identifiers}
}
\value{
the object created by caret::knn3Train
}
\description{
KNN method
}
