% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLmethods.R
\name{peppuR_rf}
\alias{peppuR_rf}
\title{Random Forest method}
\usage{
peppuR_rf(X, data, outcome_cname, ...)
}
\arguments{
\item{X}{- n-by-p data.frame containing all of the 'covariates' that will be used to predict class of y}

\item{data}{- a data.frame with columns:
 y - n vector of class identifiers for each for of X
ID - n vector of patient IDs}

\item{...}{- additional arguements passed to ranger(...)}
}
\value{
the object created by ranger::randomforest
}
\description{
Random Forest method
}
