% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.featSelect}
\alias{plot.featSelect}
\title{plot.featSelect
Plots an object of class featSelect}
\usage{
plot.featSelect(featSelect_object, order_plot = TRUE, ...)
}
\arguments{
\item{featSelect_object}{an object of the class 'featSelect'}

\item{order_plot}{logical indicator for plotting order plots}
}
\value{
plots ggplot2 object
}
\description{
For plotting an S3 object of type 'featSelect'
}
