% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot-peppuR}
\alias{plot-peppuR}
\title{plots MLinput, featSelect objects}
\description{
plots MLinput, featSelect objects
}
\details{
Various further arguments can be specified depending on the class of the object being plotted.
}
