% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MLSelection.R
\name{plot.mlSelect}
\alias{plot.mlSelect}
\title{Plots an object of class mlSelect}
\usage{
plot.mlSelect(mlSelect_object, roc_curves = TRUE, time_chart = TRUE,
  ...)
}
\arguments{
\item{mlSelect_object}{an object of the class 'mlSelect', usually created by \code{\link{MLSelection}}.}

\item{roc_curves}{logical indicator for plotting ROC curves}

\item{time_chart}{logical indicator for plotting method run times}
}
\value{
plots ggplot2 object
}
\description{
For plotting an S3 object of type 'uniFiltRes'
}
