% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rofi.R
\name{plot.rofi}
\alias{plot.rofi}
\title{Plots an object of class featSelect}
\usage{
plot.rofi(featSelect_object, order_plot = TRUE, ...)
}
\arguments{
\item{featSelect_object}{an object of the class 'featSelect', currently created by \code{\link{rofi}}.}

\item{order_plot}{logical indicator for plotting order plots}
}
\value{
plots ggplot2 object
}
\description{
For plotting an S3 object of type 'featSelect'
}
