% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_uniFiltRes.R
\name{plot.uniFiltRes}
\alias{plot.uniFiltRes}
\title{Plots an object of class uniFiltRes}
\usage{
plot.uniFiltRes(uniFilt_results, pval_threshold = NULL)
}
\arguments{
\item{pval_threshold}{is the cutoff value for pvalues, can be a single value or a vector of distinct pvalues}

\item{uniFiltRes}{an object of the class 'uniFiltRes', usually created by \code{\link{univariate_feature_selection}}.}
}
\value{
plots ggplot2 object
}
\description{
For plotting an S3 object of type 'uniFiltRes'
}
\examples{
dontrun{
library(peppuR)
library(missForest)
library(mice)

data('single_source')
data('multi_source')

x_multi = multi_source$X
y_multi = multi_source$Y

x_single = single_source$X
y_single = single_source$Y

sample_cname = 'ID'
outcome_cname = 'Group'
pair_cname = 'paircol'

result = as.MLinput(x = x_single, y = y_single, categorical_features = T , sample_cname = sample_cname, outcome_cname = outcome_cname, pair_cname = pair_cname)
result2 = as.MLinput(x = x_multi, y = y_multi, categorical_features = T, sample_cname = sample_cname, outcome_cname = outcome_cname, pair_cname = pair_cname)

imputed_res = impute_missing(result, method = 'randomforest')
imputed_res2 = impute_missing(result2, method = 'randomforest')

ufs_result = univariate_feature_selection(imputed_res)
ufs_result2 = univariate_feature_selection(imputed_res2)

plot(ufs_result)
plot(ufs_result2)

}

}
