% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_feature_selection.R
\name{univariate_feature_selection}
\alias{univariate_feature_selection}
\title{Computes features p-values}
\usage{
univariate_feature_selection(data_object)
}
\arguments{
\item{data_object}{argument is the output produced by as.MLinput function, which contains a single x data frame or a list of x data frames, a y data frames and attributes}
}
\description{
This function computes p-values for features of a single data frame or a list of data frames using, glm, lmer and glmer functions
}
\details{

}
\examples{
dontrun{
library(peppuR)
library(missForest)
library(mice)

data("single_source")
data("multi_source")

x_multi = multi_source$X
y_multi = multi_source$Y

x_single = single_source$X
y_single = single_source$Y

sample_cname = "ID"
outcome_cname = "Group"
pair_cname = "paircol"

result = as.MLinput(X = x_single, Y = y_single, categorical_features = T , sample_cname = sample_cname, outcome_cname = outcome_cname, pair_cname = pair_cname)
result2 = as.MLinput(X = x_multi, Y = y_multi, categorical_features = T, sample_cname = sample_cname, outcome_cname = outcome_cname, pair_cname = pair_cname)

imputed_res = impute_missing(result, method = "randomforest")
imputed_res2 = impute_missing(result2, method = "randomforest")

ufs_result = univariate_feature_selection(imputed_res)
ufs_result2 = univariate_feature_selection(imputed_res2)

}

}
