% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceFilter.R
\name{varianceFilter}
\alias{varianceFilter}
\title{Applies near Zero Variance function to data or data partitions}
\usage{
varianceFilter(data_object, ...)
}
\arguments{
\item{data_object}{argument is the output produced by as.MLinput, which contains a single x data frame or a list of x data frames, a y data frames and attributes}
}
\description{
This function Applies near Zero Variance function to data or data partitions
}
\details{

}
