% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.MLinput}
\alias{plot.MLinput}
\title{Plots an object of class MLInput}
\usage{
plot.MLinput(MLinput_object, order_plot = TRUE, ...)
}
\arguments{
\item{MLinput_object}{an object of the class 'MLInput', usually created by \code{\link{as.MLInput}}.}
}
\value{
plots ggplot2 object
}
\description{
Plots an object of class MLInput
}
