% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MLSelection.R
\name{plot.mlSelect}
\alias{plot.mlSelect}
\title{Plots an object of class mlSelect}
\usage{
plot.mlSelect(results_object, roc_curves = TRUE, ...)
}
\arguments{
\item{uniFiltRes}{an object of the class 'uniFiltRes', usually created by \code{\link{univariate_feature_selection}}.}

\item{pval_threshold}{is the cutoff value for pvalues, can be a single value or a vector of distinct pvalues}
}
\value{
plots ggplot2 object
}
\description{
For plotting an S3 object of type 'uniFiltRes'
}
\examples{

}
