% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rofi.R
\name{rofi}
\alias{rofi}
\title{Repeated Optimized Feature Integration}
\usage{
rofi(MLinput, source_alg_pairs, nn = 1, f_prob = 0.1, nu = 1/100,
  maxIter = 2 * sum(attr(MLinput, "data_info")$number_of_features),
  conv_check = (sum(attr(MLinput, "data_info")$number_of_features) + 1),
  epsilon = 0.01, verbose = FALSE)
}
\description{
Repeated Optimized Feature Integration
}
